+++ CHANGELOG.md
@@ -9,13 +9,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add `Scene.isActorInDrawTree` method to determine if an actor is in the scene's draw tree.
+
 ### Changed
 
 ### Deprecated
 
 ### Fixed
 
-- Fixed missing `exitviewport/enterviewport` events on Actors.on/once/off signatures. ([#978](https://github.com/excaliburjs/Excalibur/issues/978))
+- Fixed missing `exitviewport/enterviewport` events on Actors.on/once/off signatures ([#978](https://github.com/excaliburjs/Excalibur/issues/978))
+- Fix issue where Actors would not be properly added to a scene if they were removed from that scene during the same frame ([#979](https://github.com/excaliburjs/Excalibur/issues/979))
 
 <!--------------------------------- DO NOT EDIT BELOW THIS LINE --------------------------------->